chore: start cleaning up token scripts
