+++ packages/tokens-studio-poc/tasks/lib/debug.js
@@ -1,0 +1,6 @@
+export function debug(msg) {
+  const debugIt = true;
+  if (debugIt) {
+    console.log(msg.toString() || msg);
+  }
+}
